requirements.txt
alembic==1.13.2
Authlib==1.3.2
beautifulsoup4==4.12.3
cloudscraper==1.2.71
Flask==3.1.0
Flask_Bcrypt==1.0.1
Flask_Login==0.6.3
flask_mail==0.10.0
Flask_Migrate==4.0.7
flask_sqlalchemy==3.1.1
google_api_python_client==2.145.0
google_auth_oauthlib==1.2.1
itsdangerous==2.2.0
lxml==4.9.2
protobuf==5.28.3
python-dotenv==1.0.1
Requests==2.32.3
SQLAlchemy==2.0.34
Werkzeug==3.1.3
